+++ src/script/config.js
@@ -1,0 +1,18 @@
+export const DRAWER_TRANSITION_TIME = 500;
+
+//Swiper configurations
+export const SLIDES_LOOP = true;
+export const SLIDES_PER_VIEW = 1;
+export const SLIDES_GAP = 20;
+export const SLIDES_TRANSITION_TIME = 750;
+export const SLIDES_AUTOPLAY_DELAY = 3000;
+export const SLIDES_AUTOPLAY_INTERACTION_DISABILITY = false;
+export const SLIDES_ACCESSIBILITY = true;
+export const SLIDES_PAGINATION_CLICKABILE = true;
+export const SLIDES_KEYBOARD_CONTROL = true;
+export const SLIDES_ONLY_IN_VIEWPORT = true;
+export const SLIDES_NEXT_NAVIGATION_ELEMENT = 'btn--next';
+export const SLIDES_PREVIOUS_NAVIGATION_ELEMENT = 'btn--prev';
+export const SLIDES_NAVIGATION_EVENT = 'keydown';
+export const SLIDES_NEXT_KEY = 'ArrowRight';
+export const SLIDES_PREVIOUS_KEY = 'ArrowLeft';
