+++ src/styles/abstracts/_index.scss
@@ -1,3 +1,3 @@
-@forward 'variables';
 @forward 'functions';
+@forward 'variables';
 @forward 'mixins';
