+++ src/styles/components/_buttons.scss
@@ -14,16 +14,14 @@
         color: abstracts.$white;
         &:hover,
         &:focus-visible {
-            box-shadow: abstracts.pxToRem(4px) abstracts.pxToRem(14px)
-                abstracts.pxToRem(24px) rgba(0, 0, 0, 0.2);
+            box-shadow: abstracts.$box-shadow;
         }
     }
     &--secondary {
         border: abstracts.pxToRem(1px) solid abstracts.$light-grey;
         &:hover,
         &:focus-visible {
-            box-shadow: abstracts.pxToRem(4px) abstracts.pxToRem(14px)
-                abstracts.pxToRem(24px) rgba(0, 0, 0, 0.2);
+            box-shadow: abstracts.$box-shadow;
         }
     }
